+++ Assets/Fonts/Pretendard-1.3.9/web/static/pretendard-subset.css.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b5ee3bba82995bc48ba40d6b6b4b0e65
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
