+++ Assets/Fonts/Pretendard-1.3.9/web/static/pretendard.css.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 81d97e8652fec7f48abdd2756e922e90
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
