+++ Assets/Fonts/Pretendard-1.3.9/web/variable/pretendardvariable.css.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a5502ef9bb9a92040ae2e25288ecc8e2
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
