+++ Assets/Scripts/Common/ScoreResult.cs
@@ -1,26 +1,30 @@
 using System;
 using System.Collections.Generic;
-using MCRGame.Common;
 
-[Serializable]
-public class ScoreResult
+
+namespace MCRGame.Common
 {
-    public int total_score { get; set; }
-    public List<YakuScore> yaku_score_list;
-
-    public override string ToString()
+    [Serializable]
+    public class ScoreResult
     {
-        return $"Total: {total_score}, yaku_score_list: {string.Join(",", yaku_score_list)}";
+        public int total_score { get; set; }
+        public List<YakuScore> yaku_score_list;
+
+        public override string ToString()
+        {
+            return $"Total: {total_score}, yaku_score_list: {string.Join(",", yaku_score_list)}";
+        }
     }
-}
 
-public class YakuScore  
-{
-    public Yaku YakuId { get; set; }
-    public int Score { get; set; }
+    public class YakuScore
+    {
+        public Yaku YakuId { get; set; }
+        public int Score { get; set; }
 
-    public YakuScore(int yid, int score){
-        this.YakuId = (Yaku)yid;
-        this.Score = score;
+        public YakuScore(int yid, int score)
+        {
+            this.YakuId = (Yaku)yid;
+            this.Score = score;
+        }
     }
 }
