+++ Assets/Scripts/Tester.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3629edf0a7796114ab6c9886fe52f494
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
