+++ Assets/Scripts/UI/GamePage/ScorePopupManager.cs
@@ -1,6 +1,5 @@
 using UnityEngine;
-using UnityEngine.UI;
-using System.Collections.Generic;
+using MCRGame.Common;
 
 namespace MCRGame.UI
 {
